Maintain Rust 1.15 compatibility.
