+++ src/path.rs
@@ -418,39 +418,6 @@ pub mod parsing {
             })
         }
 
-        /// Parse a `Path` in mod style, while accepting keywords.
-        ///
-        /// This function is only available within `syn` to parse meta items.
-        ///
-        /// *This function is available if Syn is built with the `"parsing"`
-        /// feature.*
-        pub(crate) fn parse_meta(input: ParseStream) -> Result<Self> {
-            Ok(Path {
-                leading_colon: input.parse()?,
-                segments: {
-                    let mut segments = Punctuated::new();
-                    loop {
-                        if !Ident::peek_any(input) {
-                            break;
-                        }
-                        let ident = Ident::parse_any(input)?;
-                        segments.push_value(PathSegment::from(ident));
-                        if !input.peek(Token![::]) {
-                            break;
-                        }
-                        let punct = input.parse()?;
-                        segments.push_punct(punct);
-                    }
-                    if segments.is_empty() {
-                        return Err(input.error("expected path"));
-                    } else if segments.trailing_punct() {
-                        return Err(input.error("expected path segment"));
-                    }
-                    segments
-                },
-            })
-        }
-
         /// Determines whether this is a path of length 1 equal to the given
         /// ident.
         ///
